MediaPlayerControls exported.
