+++ example/ExampleMediaPlayer.js
@@ -2,7 +2,7 @@
 
 // eslint-disable-next-line no-unused-vars
 function ExampleMediaPlayer() {
-  return React.createElement(MediaPlayer, {
+  return React.createElement(MediaPlayer.MediaPlayerControls, {
     showVideo: true,
     controls: [
       'spacer',
